Rename Sentiment_Analysis_CodePolarity.R to R code for Polarity.R
This is the code for doing sentiment analysis with polarity. Twitter data collected through twitter API contains unwanted symbols,numbers and many other characters that we do not need for calculating sentiment analysis. Therefore we need to clean the data set first.

Each word from individual tweet is compare with positive and negative words of a dictionary. If the first words is a positive one then it stored in pos otherwise in neg variable. the next positive or negative word is added with pos or neg counter .

By Polarity Categories :( Method- Learning Based) we first classify the polarity ( positive,negative and neutral) then using ggplot2 we create a bar chart to represent the polarity categories. The highest  number of polarity represents the user sentiment about a particular subject.